+++ src/contribution/contribution.entity.ts
@@ -71,7 +71,10 @@ export class DbContribution extends BaseEntity {
   public updated_at?: Date;
 
   @ApiHideProperty()
-  @DeleteDateColumn({ type: "timestamp without time zone" })
+  @DeleteDateColumn({
+    type: "timestamp without time zone",
+    select: false,
+  })
   public deleted_at?: Date;
 
   @ApiModelProperty({