+++ src/repo/entities/repo.entity.ts
@@ -126,11 +126,15 @@ export class DbRepo extends BaseEntity {
   @Column({
     type: "timestamp without time zone",
     default: () => "to_timestamp(0)",
+    select: false,
   })
   public last_fetched_contributors_at?: Date;
 
   @ApiHideProperty()
-  @DeleteDateColumn({ type: "timestamp without time zone" })
+  @DeleteDateColumn({
+    type: "timestamp without time zone",
+    select: false,
+  })
   public deleted_at?: Date;
 
   @ApiModelProperty({